requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Optional dependencies for enhanced functionality
pandas>=2.0.0          # For data analysis and export
tabulate>=0.9.0        # For formatted table output
colorama>=0.4.6        # For colored terminal output

# Development dependencies
pytest>=7.0.0          # For testing
black>=23.0.0          # For code formatting
flake8>=6.0.0          # For linting